Use size_t for indexing and get rid of all warnings

+++ src/SparseMatrix/SparseMatrix.h
@@ -26,8 +26,8 @@
 
 				// === CREATION ==============================================
 
-				SparseMatrix(int n); // square matrix n×n
-				SparseMatrix(int rows, int columns); // general matrix
+				SparseMatrix(size_t n); // square matrix n×n
+				SparseMatrix(size_t rows, size_t columns); // general matrix
 
 				SparseMatrix(const SparseMatrix<T> & m); // copy constructor
 				SparseMatrix<T> & operator = (const SparseMatrix<T> & m);
@@ -37,14 +37,14 @@
 
 				// === GETTERS / SETTERS ==============================================
 
-				int getRowCount(void) const;
-				int getColumnCount(void) const;
+                size_t getRowCount(void) const;
+                size_t getColumnCount(void) const;
 
 
 				// === VALUES ==============================================
 
-				T get(int row, int col) const;
-				SparseMatrix & set(T val, int row, int col);
+				T get(size_t row, size_t col) const;
+				SparseMatrix & set(T val, size_t row, size_t col);
 
 
 				// === OPERATIONS ==============================================
@@ -76,34 +76,34 @@
 
 			protected:
 
-				int m, n;
+                size_t m, n;
 
 				std::vector<T> * vals;
-				std::vector<int> * rows, * cols;
+				std::vector<size_t> * rows, * cols;
 
 
 				// === HELPERS / VALIDATORS ==============================================
 
-				void construct(int m, int n);
+				void construct(size_t m, size_t n);
 				void destruct(void);
 				void deepCopy(const SparseMatrix<T> & m);
-				void validateCoordinates(int row, int col) const;
-				void insert(int index, int row, int col, T val);
-				void remove(int index, int row);
+				void validateCoordinates(size_t row, size_t col) const;
+				void insert(size_t index, size_t row, size_t col, T val);
+				void remove(size_t index, size_t row);
 
 		};
 
     // === CREATION ==============================================
 
     template<typename T>
-    SparseMatrix<T>::SparseMatrix(int n)
+    SparseMatrix<T>::SparseMatrix(size_t n)
     {
         this->construct(n, n);
     }
 
 
     template<typename T>
-    SparseMatrix<T>::SparseMatrix(int rows, int columns)
+    SparseMatrix<T>::SparseMatrix(size_t rows, size_t columns)
     {
         this->construct(rows, columns);
     }
@@ -133,10 +133,10 @@
     {
         this->m = matrix.m;
         this->n = matrix.n;
-        this->rows = new std::vector<int>(*(matrix.rows));
+        this->rows = new std::vector<size_t>(*(matrix.rows));
 
         if (matrix.vals != NULL) {
-            this->cols = new std::vector<int>(*(matrix.cols));
+            this->cols = new std::vector<size_t>(*(matrix.cols));
             this->vals = new std::vector<T>(*(matrix.vals));
         }
     }
@@ -150,7 +150,7 @@
 
 
     template<typename T>
-    void SparseMatrix<T>::construct(int rows, int columns)
+    void SparseMatrix<T>::construct(size_t rows, size_t columns)
     {
         if (rows < 1 || columns < 1) {
             throw InvalidDimensionsException("Matrix dimensions cannot be zero or negative.");
@@ -161,7 +161,7 @@
 
         this->vals = NULL;
         this->cols = NULL;
-        this->rows = new std::vector<int>(rows + 1, 1);
+        this->rows = new std::vector<size_t>(rows + 1, 1);
     }
 
 
@@ -180,14 +180,14 @@
     // === GETTERS / SETTERS ==============================================
 
     template<typename T>
-    int SparseMatrix<T>::getRowCount(void) const
+    size_t SparseMatrix<T>::getRowCount(void) const
     {
         return this->m;
     }
 
 
     template<typename T>
-    int SparseMatrix<T>::getColumnCount(void) const
+    size_t SparseMatrix<T>::getColumnCount(void) const
     {
         return this->n;
     }
@@ -196,13 +196,13 @@
     // === VALUES ==============================================
 
     template<typename T>
-    T SparseMatrix<T>::get(int row, int col) const
+    T SparseMatrix<T>::get(size_t row, size_t col) const
     {
         this->validateCoordinates(row, col);
 
-        int currCol;
+        size_t currCol;
 
-        for (int pos = (*(this->rows))[row - 1] - 1; pos < (*(this->rows))[row] - 1; ++pos) {
+        for (size_t pos = (*(this->rows))[row - 1] - 1; pos < (*(this->rows))[row] - 1; ++pos) {
             currCol = (*(this->cols))[pos];
 
             if (currCol == col) {
@@ -218,12 +218,12 @@
 
 
     template<typename T>
-    SparseMatrix<T> & SparseMatrix<T>::set(T val, int row, int col)
+    SparseMatrix<T> & SparseMatrix<T>::set(T val, size_t row, size_t col)
     {
         this->validateCoordinates(row, col);
 
-        int pos = (*(this->rows))[row - 1] - 1;
-        int currCol = 0;
+        size_t pos = (*(this->rows))[row - 1] - 1;
+        size_t currCol = 0;
 
         for (; pos < (*(this->rows))[row] - 1; pos++) {
             currCol = (*(this->cols))[pos];
@@ -261,9 +261,9 @@
         std::vector<T> result(this->m, T());
 
         if (this->vals != NULL) { // only if any value set
-            for (int i = 0; i < this->m; i++) {
+            for (size_t i = 0; i < this->m; i++) {
                 T sum = T();
-                for (int j = (*(this->rows))[i]; j < (*(this->rows))[i + 1]; j++) {
+                for (size_t j = (*(this->rows))[i]; j < (*(this->rows))[i + 1]; j++) {
                     sum = sum + (*(this->vals))[j - 1] * x[(*(this->cols))[j - 1] - 1];
                 }
 
@@ -380,7 +380,7 @@
     // === HELPERS / VALIDATORS ==============================================
 
     template<typename T>
-    void SparseMatrix<T>::validateCoordinates(int row, int col) const
+    void SparseMatrix<T>::validateCoordinates(size_t row, size_t col) const
     {
         if (row < 1 || col < 1 || row > this->m || col > this->n) {
             throw InvalidCoordinatesException("Coordinates out of range.");
@@ -389,30 +389,30 @@
 
 
     template<typename T>
-    void SparseMatrix<T>::insert(int index, int row, int col, T val)
+    void SparseMatrix<T>::insert(size_t index, size_t row, size_t col, T val)
     {
         if (this->vals == NULL) {
             this->vals = new std::vector<T>(1, val);
-            this->cols = new std::vector<int>(1, col);
+            this->cols = new std::vector<size_t>(1, col);
 
         } else {
             this->vals->insert(this->vals->begin() + index, val);
             this->cols->insert(this->cols->begin() + index, col);
         }
 
-        for (int i = row; i <= this->m; i++) {
+        for (size_t i = row; i <= this->m; i++) {
             (*(this->rows))[i] += 1;
         }
     }
 
 
     template<typename T>
-    void SparseMatrix<T>::remove(int index, int row)
+    void SparseMatrix<T>::remove(size_t index, size_t row)
     {
         this->vals->erase(this->vals->begin() + index);
         this->cols->erase(this->cols->begin() + index);
 
-        for (int i = row; i <= this->m; i++) {
+        for (size_t i = row; i <= this->m; i++) {
             (*(this->rows))[i] -= 1;
         }
     }
