+++ tests/cases/crs-format.cpp
@@ -35,14 +35,14 @@ void testInternalStorage(void)
 		.set(3, 3, 3)
 		.set(2, 3, 4);
 
-	std::vector<int> rowPointers1;
+	std::vector<size_t> rowPointers1;
 	rowPointers1.push_back(1);
 	rowPointers1.push_back(4);
 	rowPointers1.push_back(6);
 	rowPointers1.push_back(8);
-	assertEquals<std::vector<int> >(rowPointers1, *(m1.getRowPointers()), "Incorrect internal row pointers");
+	assertEquals<std::vector<size_t> >(rowPointers1, *(m1.getRowPointers()), "Incorrect internal row pointers");
 
-	std::vector<int> columnPointers1;
+	std::vector<size_t> columnPointers1;
 	columnPointers1.push_back(1);
 	columnPointers1.push_back(3);
 	columnPointers1.push_back(4);
@@ -50,7 +50,7 @@ void testInternalStorage(void)
 	columnPointers1.push_back(2);
 	columnPointers1.push_back(3);
 	columnPointers1.push_back(4);
-	assertEquals<std::vector<int> >(columnPointers1, *(m1.getColumnPointers()), "Incorrect internal column pointers");
+	assertEquals<std::vector<size_t> >(columnPointers1, *(m1.getColumnPointers()), "Incorrect internal column pointers");
 
 	std::vector<int> values1;
 	values1.push_back(1);
@@ -82,20 +82,20 @@ void testInternalStorage(void)
 		. set(1, 3, 2)
 		. set(4, 3, 4);
 
-	std::vector<int> rowPointers2;
+	std::vector<size_t> rowPointers2;
 	rowPointers2.push_back(1);
 	rowPointers2.push_back(3);
 	rowPointers2.push_back(3);
 	rowPointers2.push_back(6);
-	assertEquals<std::vector<int> >(rowPointers2, *(m2.getRowPointers()), "Incorrect internal row pointers");
+	assertEquals<std::vector<size_t> >(rowPointers2, *(m2.getRowPointers()), "Incorrect internal row pointers");
 
-	std::vector<int> columnPointers2;
+	std::vector<size_t> columnPointers2;
 	columnPointers2.push_back(1);
 	columnPointers2.push_back(4);
 	columnPointers2.push_back(1);
 	columnPointers2.push_back(2);
 	columnPointers2.push_back(4);
-	assertEquals<std::vector<int> >(columnPointers2, *(m2.getColumnPointers()), "Incorrect internal column pointers");
+	assertEquals<std::vector<size_t> >(columnPointers2, *(m2.getColumnPointers()), "Incorrect internal column pointers");
 
 	std::vector<int> values2;
 	values2.push_back(10);
@@ -118,21 +118,21 @@ void testInternalStorage(void)
 	SparseMatrixMock<int> m3 = m2;
 	m3.set(5, 2, 2);
 
-	std::vector<int> rowPointers3;
+	std::vector<size_t> rowPointers3;
 	rowPointers3.push_back(1);
 	rowPointers3.push_back(3);
 	rowPointers3.push_back(4);
 	rowPointers3.push_back(7);
-	assertEquals<std::vector<int> >(rowPointers3, *(m3.getRowPointers()), "Incorrect internal row pointers");
+	assertEquals<std::vector<size_t> >(rowPointers3, *(m3.getRowPointers()), "Incorrect internal row pointers");
 
-	std::vector<int> columnPointers3;
+	std::vector<size_t> columnPointers3;
 	columnPointers3.push_back(1);
 	columnPointers3.push_back(4);
 	columnPointers3.push_back(2);
 	columnPointers3.push_back(1);
 	columnPointers3.push_back(2);
 	columnPointers3.push_back(4);
-	assertEquals<std::vector<int> >(columnPointers3, *(m3.getColumnPointers()), "Incorrect internal column pointers");
+	assertEquals<std::vector<size_t> >(columnPointers3, *(m3.getColumnPointers()), "Incorrect internal column pointers");
 
 	std::vector<int> values3;
 	values3.push_back(10);
@@ -155,8 +155,8 @@ void testInternalStorage(void)
 	SparseMatrixMock<int> m4 = m3;
 	m4.set(0, 2, 2);
 
-	assertEquals<std::vector<int> >(rowPointers2, *(m4.getRowPointers()), "Incorrect internal row pointers");
-	assertEquals<std::vector<int> >(columnPointers2, *(m4.getColumnPointers()), "Incorrect internal column pointers");
+	assertEquals<std::vector<size_t> >(rowPointers2, *(m4.getRowPointers()), "Incorrect internal row pointers");
+	assertEquals<std::vector<size_t> >(columnPointers2, *(m4.getColumnPointers()), "Incorrect internal column pointers");
 	assertEquals<std::vector<int> >(values2, *(m4.getValues()), "Incorrect internal values storage");
 
 	std::cout << " OK" << std::endl;
